refactor: trackers are generic
In anticipation of adding other kinds of trackers,
we need to remove the more "hard coded" Kubernetes
logic and have it exposed via a generic tracker
class. This first PR does that. For orchestration
outside of Kubernetes, the idea would be that we do
not need the operator itself.
