+++ .github/workflows/e2e-test.yaml
@@ -49,7 +49,7 @@ jobs:
           sudo rm -rf /usr/share/dotnet
           sudo rm -rf /usr/local/lib/android
           sudo rm -rf /opt/ghc
-          
+
     - name: Load Docker images
       run: |
         ls /tmp/*.tar.gz
@@ -62,11 +62,11 @@ jobs:
         cluster_name: kind
         kubectl_version: v1.28.2
         version: v0.20.0
-        config: ./.github/test-kind-config.yaml 
+        config: ./.github/test-kind-config.yaml
     - name: Install Helm
       run: |
         export HELM_EXPERIMENTAL_OCI=1
-        curl https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3 | bash        
+        curl https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3 | bash
     - name: Deploy State Machine Operator
       run: make test-deploy-ci && sleep 10
     - name: Test Workflow
