+++ setup.cfg
@@ -12,4 +12,5 @@ per-file-ignores =
     python/state_machine_operator/config/__init__.py:F401
     python/state_machine_operator/machine/__init__.py:F401
     python/state_machine_operator/tracker/__init__.py:F401
+    python/state_machine_operator/tracker/kubernetes/__init__.py:F401
     python/state_machine_operator/manager/__init__.py:F401
